fix: 프로필 수정 시 프로필 이미지 null 처리 안한 부분 수정

See merge request pttrack/class_01/web_project_i/team04/team04-backend!69

+++ src/main/java/com/elice/team04backend/service/impl/UserServiceImpl.java
@@ -45,7 +45,9 @@ public class UserServiceImpl implements UserService {
 
         try {
             if (!profileImage.isEmpty()) {
-                firebaseStorageService.deleteImage(originalProfileImageUrl);
+                if(originalProfileImageUrl != null) {
+                    firebaseStorageService.deleteImage(originalProfileImageUrl);
+                }
                 updatedProfileImageUrl = firebaseStorageService.uploadImage(profileImage);
             }
         } catch (IOException e) {
